#27 - 로그인 컨트롤러 테스트 + 시큐리티 설정을 컨트롤러 테스트에 반영
스프링 시큐리티가 자동으로 생성해준 `/login` 컨트롤러 테스트
그리고 컨트롤러가 시큐리티 설정을 부를 수 있도록 `@Import` 추가
