+++ part2/fastcampus-project-board/src/test/java/com/fastcampus/projectboard/controller/AuthControllerTest.java
@@ -1,0 +1,40 @@
+package com.fastcampus.projectboard.controller;
+
+import com.fastcampus.projectboard.config.SecurityConfig;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.context.annotation.Import;
+import org.springframework.http.MediaType;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
+
+@DisplayName("View 컨트롤러 - 인증")
+@Import(SecurityConfig.class)
+@WebMvcTest
+public class AuthControllerTest {
+
+    private final MockMvc mvc;
+
+    public AuthControllerTest(@Autowired MockMvc mvc) {
+        this.mvc = mvc;
+    }
+
+    @DisplayName("[view][GET] 게시글 리스트 (게시판) 페이지 - 정상 호출")
+    @Test
+    void articleListViewTest() throws Exception {
+        mvc.perform(MockMvcRequestBuilders.get("/articles"))
+                .andExpect(status().isOk())
+                .andExpect(content().contentTypeCompatibleWith(MediaType.TEXT_HTML)) // view 이기 때문에 응답 타입은 text/html
+                .andExpect(view().name("articles/index"))
+                .andExpect(model().attributeExists("articles")) // model 에 해당 이름으로 넘긴 데이터가 있는지 체크
+                .andDo(print());
+
+    }
+
+}
